simplify retry
