+++ docker-compose.yml
@@ -115,7 +115,7 @@ services:
   database:
     container_name: stac-db
     platform: linux/amd64
-    image: ghcr.io/stac-utils/pgstac:v0.6.11
+    image: ghcr.io/stac-utils/pgstac:v0.7.4
     environment:
       - POSTGRES_USER=username
       - POSTGRES_PASSWORD=password
