+++ titiler/pgstac/dependencies.py
@@ -7,6 +7,7 @@ import pystac
 from cachetools import TTLCache, cached
 from cachetools.keys import hashkey
 from fastapi import HTTPException, Path, Query
+from psycopg import errors as pgErrors
 from psycopg.rows import dict_row
 from psycopg_pool import ConnectionPool
 from starlette.requests import Request
@@ -14,6 +15,7 @@ from starlette.requests import Request
 from titiler.core.dependencies import DefaultDependency
 from titiler.pgstac import model
 from titiler.pgstac.settings import CacheSettings
+from titiler.pgstac.utils import retry
 
 cache_config = CacheSettings()
 
@@ -80,6 +82,7 @@ class PgSTACParams(DefaultDependency):
     TTLCache(maxsize=cache_config.maxsize, ttl=cache_config.ttl),
     key=lambda pool, collection, item: hashkey(collection, item),
 )
+@retry(tries=3, exceptions=pgErrors.OperationalError)
 def get_stac_item(pool: ConnectionPool, collection: str, item: str) -> pystac.Item:
     """Get STAC Item from PGStac."""
     search = model.PgSTACSearch(ids=[item], collections=[collection])
